+++ src/publisher/index.ts
@@ -14,6 +14,7 @@
  * limitations under the License.
  */
 
+import {promisify} from '@google-cloud/promisify';
 import * as extend from 'extend';
 import {CallOptions} from 'google-gax';
 import {isSpanContextValid, Span} from '@opentelemetry/api';
@@ -119,7 +120,7 @@ export class Publisher {
             const flushResolver = () => {
               resolve();
 
-              // flush() maybe called more than once, so remove these
+              // flush() may be called more than once, so remove these
               // event listeners after we've completed flush().
               q.removeListener('drain', flushResolver);
             };
@@ -128,7 +129,9 @@ export class Publisher {
       )
     );
 
-    const allPublishes = Promise.all(toDrain.map(q => q.publish.bind(q)()));
+    const allPublishes = Promise.all(
+      toDrain.map(q => promisify(q.publishDrain).bind(q)())
+    );
 
     allPublishes
       .then(() => allDrains)